requirements.txt
aioredis==2.0.1
amqp==5.1.1
anyio==3.6.1
astroid==2.12.9
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.8.0
celery==5.2.7
certifi==2022.9.14
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.4
dill==0.3.5.1
distlib==0.3.6
dnspython==2.2.1
email-validator==1.2.1
espeakng==1.0.2
fastapi==0.85.0
filelock==3.8.0
flake8==5.0.4
flake8-bugbear==22.9.11
flake8-docstrings==1.6.0
flake8-import-order==0.18.1
flake8-requirements==1.7.1
gTTS==2.3.0
h11==0.13.0
httptools==0.5.0
identify==2.5.5
idna==3.4
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
nodeenv==1.7.0
orjson==3.8.0
packaging==21.3
pathspec==0.10.1
pep8-naming==0.13.2
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
prompt-toolkit==3.0.31
py==1.11.0
pycodestyle==2.9.1
pydantic==1.10.2
pydocstyle==6.1.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.19.0
pytest-cov==3.0.0
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
starlette==0.20.4
tesserocr==2.5.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
ujson==5.5.0
urllib3==1.26.12
urlopen==1.0.0
uvicorn==0.18.3
uvloop==0.17.0
vine==5.0.0
virtualenv==20.16.5
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
python-docx==0.8.11
PyPDF2==2.11.1
pandas==1.5.1
